# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: AviaOutdoor API
  version: 1.0.0
servers:
  - url: http://localhost:4000
paths:
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: OK
  /api/auth/login:
    post:
      summary: Login and get JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Login success
        '401':
          description: Invalid username or password
        '429':
          description: Too many login attempts
  /api/auth/me:
    get:
      summary: Get current authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
  /api/schema:
    get:
      summary: Database schema summary
      responses:
        '200':
          description: OK
  /api/categories:
    get:
      summary: List categories
      responses:
        '200':
          description: OK
    post:
      summary: Create category
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Created
  /api/customers:
    get:
      summary: List/search customers
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: q
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
  /api/categories/{name}:
    delete:
      summary: Delete category
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
  /api/items:
    get:
      summary: List items
      responses:
        '200':
          description: OK
    post:
      summary: Create item
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Created
  /api/items/{id}:
    patch:
      summary: Update item
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Updated
    delete:
      summary: Delete item
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
  /api/rentals:
    get:
      summary: List rentals
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
    post:
      summary: Create rental
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Created
  /api/returns:
    get:
      summary: List return records
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
    post:
      summary: Process return
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Processed
  /api/users:
    get:
      summary: List users (admin only)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
    post:
      summary: Create user (admin only)
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Created
  /api/users/{id}:
    patch:
      summary: Update user profile by admin
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Updated
  /api/users/me/password:
    patch:
      summary: Change current user password
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Updated
  /api/users/{id}/password:
    patch:
      summary: Reset user password by admin
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Updated
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT